requirements.txt
# Core scraping and data handling
telethon
pandas
python-dotenv
oracledb
fastapi
uvicorn
pydantic
requests
aiohttp
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
alembic
python-multipart
httpx
pytz
lru-dict
prometheus-fastapi-instrumentator
torch
opencv-python
Pillow

# Database & ORM
psycopg2-binary
SQLAlchemy
ultralytics

# Data transformation
# dbt is split into core + adapter
dbt-core
dbt-postgres

# API
fastapi
uvicorn[standard]

# Orchestration
Dagster

# Computer Vision / Object Detection
opencv-python-headless
pillow
# Torch is large; you may prefer installing with CUDA/cuDNN variants manually
torch
torchvision

# Dev / Testing
black
ruff
pytest

pydantic
pydantic-settings

# Utilities
python-slugify
tqdm
